+++ CHANGELOG.md
@@ -1,12 +1,19 @@
 # Changelog
 
+## 1.8.14
+
+- editor
+  - add `mod+d` and `mod+shift+d` for duplicating widgets
+  - use a temporary clipboard when duplicating widgets
+
+
 ## 1.8.13
 
 - widgets
   - file: show save icon when mode is set to save
   - file: center icon when hidePath is set to true
   - sliders/pads: apply `spring` property dynamically
-  
+
 - ui
   - add `alt`+`wheel` for local zoom centered on cursor